// src/app/models/question.ts
export class QuestionBase<T> {
  value: T | undefined;
  key: string;
  label: string;
  display:boolean;
  required: boolean;
  order: number;
  controlType: string;
  type: string;
  options: { key: string, value: string, label: string, inputModel: string }[];

  constructor(options: {
    value?: T;
    key?: string;
    label?: string;
    required?: boolean;
    display?:boolean;
    order?: number;
    controlType?: string;
    type?: string;
    options?: { key: string, value: string, label: string, inputModel: string }[];
  } = {}) {
    this.value = options.value;
    this.key = options.key || '';
    this.label = options.label || '';
    this.display = !!options.display;
    this.required = !!options.required;
    this.order = options.order === undefined ? 1 : options.order;
    this.controlType = options.controlType || '';
    this.type = options.type || '';
    this.options = options.options || [];
  }
}